requirements.txt
anthropic>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
logging>=0.5.1.2
beautifulsoup4>=4.12.0
markdown>=3.5.0